"Add error page for undefined routes and beforeEnter guards for donation form and mosques info pages to check authentication"
This commit adds a new route for an error page that will be displayed when the user navigates to a route that doesn't exist. It also adds beforeEnter guards to the donation form and mosques info pages to check if the user is logged in or an admin, respectively, before allowing them to access the page.